Deploy the Notebook as a Web Application
Deploy the Notebook as a Web Application (10 points)
To receive all points, you must:

Use the Voilà library to deploy your notebook as a web application. Deploy the web application locally on your computer. (5 points)

Take a screen recording or screenshots to show how the web application appears when using Voilà. Include the recording or screenshots in the README.md file for your GitHub repository. (5 points)